Get the taxonomic classification hierarchy for taxa resolved to supported authorities

Get the taxonomic classification hierarchy for taxa resolved to supported authorities

get_classification(
  taxa.clean,
  authority = NA,
  authority.id = NA,
  rank = NA,
  path = NULL
)

Arguments

taxa.clean

(character) Taxa names

authority

(character) Authority taxa.clean have been resolved to, otherwise NA. Supported authorities include: "ITIS", "WORMS", "GBIF".

authority.id

(character) ID of taxa.clean within the authority, otherwise NA

rank

(character) Rank (e.g. "Genus", "Species") of taxa.clean, otherwise NA. This is useful when taxa.clean can't be resolved to an authority and the rank must be manually defined.

path

(character) Path of the directory containing taxa_map.csv.

Value

(list) For taxa.clean resolved to a supported authority, each item in the list is a classification hierarchy (also a list), including one or more common names (only when authority is ITIS or WORMS) and authority IDs for each rank-value pair. For taxa.clean not resolved to a supported authority, each item is listed as defined in the taxa.clean, authority, and authority.id arguments.

Details

Only taxa resolved to supported authorities can be expanded into a full taxonomic classification with common names. Taxa resolved to unsupported authorities, or not resolved at all, will be listed as is defined in the taxa.clean, authority, and authority.id arguments.

Supported authorities are recognized by a controlled set of representations.

ITIS can be: "ITIS", "itis", "Integrated Taxonomic Information System", or "https://www.itis.gov/".

WORMS can be: "WORMS", "worms", "World Register of Marine Species", or "http://www.marinespecies.org/".

GBIF can be: "GBIF", "gbif", "GBIF Backbone Taxonomy", or "https://gbif.org".